+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next ##
 
+* Engine: Changed defaults and acceptable ranges for options
 * Engine: Exposed QuiescentSearchMaxDepth and PrincipalVariationMaxDepth options
 
 ## v0.12.9 ##